rust: BIT_AND/BIT_OR/BIT_XOR return BIGINT UNSIGNED, as Go infers them
Go aggregation/base_func.go:381-386 (typeInfer4BitFuncs) adds
mysql.UnsignedFlag alongside NotNullFlag, and func_bitfuncs.go appends
the fold with AppendUint64. We set only NOT_NULL, so the identity fold of
an all-NULL group printed -1 instead of 18446744073709551615.

Captured from TiDB over v BIGINT:
  values (null)         -> 18446744073709551615 | 0                    | 0
  values (-1),(-1)      -> 18446744073709551615 | 18446744073709551615 | 0
  values (3),(5),(null) -> 1 | 7 | 6
  a view over them describes as bigint(21) unsigned NO

The doc comments in both files asserted the opposite and cited a capture
for it; that capture was wrong and the comments are corrected.

hash_agg's finalizer now yields Datum::UInt to match Go's AppendUint64
and the sibling tier (tidb-exec/src/bit_agg.rs). That half is not
independently observable -- Datum equality is numeric and the downstream
conversion already normalizes to the column's signedness -- so the flag
is the load-bearing change; the mutation table in the report says so.

+++ rust/crates/tidb-executor/src/hash_agg.rs
@@ -825,11 +825,12 @@ impl Partial {
                 }
             }
             Partial::AvgReal { sum, count } => Datum::Real(sum / *count as f64),
-            // Go's result column is a SIGNED `BIGINT` holding the unsigned
-            // fold's bit pattern, which is why `BIT_AND` over an all-NULL
-            // group prints `-1` rather than `18446744073709551615`
-            // (captured from TiDB).
-            Partial::Bit { acc, .. } => Datum::Int(*acc as i64),
+            // Go `typeInfer4BitFuncs` marks the column `UnsignedFlag`, and
+            // `func_bitfuncs.go`'s `AppendFinalResult2Chunk` does
+            // `AppendUint64`, so the fold is printed as the unsigned value:
+            // `BIT_AND` over an all-NULL group is `18446744073709551615`,
+            // not `-1` (captured from TiDB).
+            Partial::Bit { acc, .. } => Datum::UInt(*acc),
             // Go: population variance divides by `count`, sample variance by
             // `count - 1` and is NULL for a single row; both are NULL for an
             // empty (or all-NULL) input.
